Rework application configuration to allow injection points in tests
