+++ wallet-provider-service/src/main/kotlin/adapter/persistence/R2dbcDatabaseExtensions.kt
@@ -15,13 +15,7 @@
  */
 package eu.europa.ec.eudi.walletprovider.adapter.persistence
 
-import org.jetbrains.exposed.v1.core.vendors.ForUpdateOption
-import org.jetbrains.exposed.v1.core.vendors.H2Dialect
-import org.jetbrains.exposed.v1.core.vendors.MariaDBDialect
-import org.jetbrains.exposed.v1.core.vendors.MysqlDialect
-import org.jetbrains.exposed.v1.core.vendors.OracleDialect
-import org.jetbrains.exposed.v1.core.vendors.PostgreSQLDialect
-import org.jetbrains.exposed.v1.core.vendors.SQLServerDialect
+import org.jetbrains.exposed.v1.core.vendors.*
 import org.jetbrains.exposed.v1.r2dbc.R2dbcDatabase
 
 val R2dbcDatabase.forUpdateOption: ForUpdateOption
